api/requirements.txt
flask==3.0.3
flask-cors==5.0.0
pillow==10.4.0
tensorflow==2.13.0
numpy==1.24.3